pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "electoral-sim"
version = "0.0.1"
description = "High-performance agent-based electoral simulation toolkit with support for 11+ countries, multiple voting systems, and advanced voter behavior models"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ayush", email = "ayush12358@users.noreply.github.com"}
]
maintainers = [
    {name = "Ayush", email = "ayush12358@users.noreply.github.com"}
]
keywords = [
    "election",
    "simulation", 
    "agent-based-modeling",
    "mesa",
    "voting",
    "electoral-systems",
    "political-science",
    "fptp",
    "proportional-representation",
    "coalition",
    "opinion-dynamics",
    "computational-social-science",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Life",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "mesa>=3.0.0",
    "mesa-frames",
    "polars>=0.20.0",
    "numpy>=1.24.0",
    "networkx>=3.0",
    "numba>=0.58.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    "streamlit>=1.30.0",
]
gpu = [
    "cupy-cuda12x>=12.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "electoral-sim[viz,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/Ayush12358/ElectoralSim"
Documentation = "https://github.com/Ayush12358/ElectoralSim#readme"
Repository = "https://github.com/Ayush12358/ElectoralSim"
Issues = "https://github.com/Ayush12358/ElectoralSim/issues"
Changelog = "https://github.com/Ayush12358/ElectoralSim/releases"

[project.scripts]
electoral-sim = "electoral_sim.core.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["electoral_sim*"]

[tool.setuptools.package-data]
electoral_sim = ["data/*.csv", "data/*.json", "py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
# Only check for errors, not style issues
select = ["E", "F"]
# Ignore: line too long, unused imports, unused vars, imports not at top, forward type hints
ignore = ["E501", "F401", "F841", "E402", "F821"]

[tool.ruff.lint.isort]
known-first-party = ["electoral_sim"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::numba.core.errors.NumbaDeprecationWarning",
]

[tool.coverage.run]
source = ["electoral_sim"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
